gimplefe: Simplify handling of identifier based binary operations
While looking into adding __ROTATE_LEFT and __ROTATE_RIGHT, I noticed
this code is just a bunch of if statments repeated. Instead we could just
use a simple lookup array to do the same thinga and it would be easier to
add to the array instead of duplicating the if sequence again.

Bootstrapped and tested on x86_64-linux-gnu.

gcc/c/ChangeLog:

	* gimple-parser.cc (gimple_binary_identifier_code): New variable.
	(c_parser_gimple_binary_expression): Use gimple_binary_identifier_code
	instead of doing if statements on the strings.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>