arm: remove iWMMX builtins support.
This is the first step of removing the various builtins for iwmmxt,
removing the builtins expansion code.  It leaves a lot of code
elsewhere, but we'll clean that up in subsequent patches.

I'm not sure why safe_vector_operand would unconditionally try to
expand to an iwmmxt instruction if passed (const_int 0).  Clearly
that's meaningless on other architectures, but perhaps this can't
happen elsewhere.  Anyway, for now, just mark this as unreachable so
that we'll know about it if it ever happens.

gcc/ChangeLog:

	* config/arm/arm-builtins.cc (enum arm_builtins): Delete iWMMX
	builtin values.
	(bdesc_2arg): Likewise.
	(bdesc_1arg): Likewise.
	(arm_init_iwmmxt_builtins): Delete.
	(arm_init_builtins): Don't call arm_init_iwmmxt_builtins.
	(safe_vector_operand): Use __builtin_unreachable instead of emitting
	an iwmmxt builtin.
	(arm_general_expand_builtin): Remove iWMMX builtins support.