gimple: Merge slow and fast bit-test switch lowering [PR117091]
PR117091 showed that bit-test switch lowering can take a lot of time.
The algorithm was O(n^2).  We therefore came up with a faster algorithm
(O(n * BITS_IN_WORD)) and made GCC choose between the slow and the fast
algorithm based on how big the switch is.

Here I combine the algorithms so that we get the results of the slower
algorithm in the faster asymptotic time.

	PR middle-end/117091

gcc/ChangeLog:

	* tree-switch-conversion.cc (bit_test_cluster::find_bit_tests_fast):
	Remove function.
	(bit_test_cluster::find_bit_tests_slow): Remove function.
	(bit_test_cluster::find_bit_tests): We don't need to decide
	between slow and fast so just put the modified (no longer) slow
	algorithm here.

Signed-off-by: Filip Kastl <fkastl@suse.cz>