expand: Remove unsignedp argument from get_compare_parts [PR118090]
While helping Eikansh with a patch to ccmp, it was noticed that the
result stored in the up pointer that gets passed to get_compare_parts
was unused on all call sites.
It was always unused since get_compare_parts was added in
r8-1717-gf580a969d7fbab. It looks it was not noticed it became unused
when rcode was set via get_compare_parts and in RTL, the signedness is
part of the comparison.

	PR middle-end/118090
gcc/ChangeLog:

	* ccmp.cc (get_compare_parts): Remove the up argument.
	(expand_ccmp_next): Update call to get_compare_parts.
	(expand_ccmp_expr_1): Likewise.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>