i386: -Wabi false positive with indirect call [PR60336]
This warning relies on the TRANSLATION_UNIT_WARN_EMPTY_P flag (set in
cxx_init_decl_processing) to decide whether we want to warn about the GCC 8
empty class parameter passing fix, but in a call through a function pointer
we don't have a translation unit and so complain for any -Wabi flag, even
now long after this was likely to be relevant.

In that situation, let's check the TU for current_function_decl instead.
And if we still can't come up with a TU, default to not warning.

	PR c++/60336

gcc/ChangeLog:

	* config/i386/i386.cc (ix86_warn_parameter_passing_abi):
	If no target, check the current TU.

gcc/testsuite/ChangeLog:

	* g++.dg/abi/pr60336-8a.C: New test.