c++: CTAD and constexpr ctor [PR115207]
Here we failed to constant-evaluate the A<int> constructor because DECL_SIZE
wasn't set on 'a' yet, so compare_address thinks we can't be sure it isn't
at the same address as 'i'.

Normally DECL_SIZE is set by build_decl calling layout_decl, but that
doesn't happen here because we don't have a type yet.  So we need to
layout_decl again after deduction.

	PR c++/115207

gcc/cp/ChangeLog:

	* decl.cc (cp_finish_decl): Call layout_decl after CTAD.

gcc/testsuite/ChangeLog:

	* g++.dg/cpp1z/class-deduction118.C: New test.