[V2][RISC-V] Trivial permutation constant derivation
This is a patch from late 2024 (just before stage1 freeze), but I never pushed
hard to the change, and thus never integrated it.

It's mostly unchanged except for updating insn in the hash table after finding
an optimizable case.  We were holding the deleted insn in the hash table rather
than the new insn.  Just something I noticed recently.

Bootstrapped and regression tested on my BPI and regression tested riscv32-elf
and riscv64-elf configurations.  We've used this since November internally, so
it's well exercised on spec as well.

gcc/
	* config.gcc (riscv): Add riscv-vect-permcost.o to extra_objs.
	* config/riscv/riscv-passes.def (pass_vector_permcost): Add new pass.
	* config/riscv/riscv-protos.h (make_pass_vector_permconst): Declare.
	* config/riscv/riscv-vect-permconst.cc: New file.
	* config/riscv/t-riscv: Add build rule for riscv-vect-permcost.o